requirements.txt
fastapi
uvicorn
google-cloud-firestore
python-dotenv
pydantic
passlib[bcrypt]
pyjwt